+++ contracts/UMB.sol
@@ -8,7 +8,6 @@ import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 import "./interfaces/SwappableToken.sol";
-import "./interfaces/BurnableToken.sol";
 import "./interfaces/MintableToken.sol";
 import "./interfaces/ISwapReceiver.sol";
 
@@ -21,7 +20,7 @@ import "./interfaces/ISwapReceiver.sol";
 ///          This token can be mint by owner eg we need UMB for auction. After that we can burn the key
 ///          so nobody can mint anymore.
 ///          It has limit for max total supply, so we need to make sure, total amount of rUMBs fit this limit.
-contract UMB is BurnableToken, MintableToken, ISwapReceiver {
+contract UMB is MintableToken, ISwapReceiver {
   using SafeMath for uint256;
 
   // ========== STATE VARIABLES ========== //
@@ -31,16 +30,16 @@ contract UMB is BurnableToken, MintableToken, ISwapReceiver {
   // ========== CONSTRUCTOR ========== //
 
   constructor (
+    address _multiSig,
     address _initialHolder,
     uint _initialBalance,
-    address _multiSig,
     uint256 _maxAllowedTotalSupply,
     string memory _name,
     string memory _symbol
   )
   Owned(_multiSig)
   ERC20(_name, _symbol)
-  MaxAllowedSupply(_maxAllowedTotalSupply) {
+  MintableToken(_maxAllowedTotalSupply) {
     if (_initialHolder != address(0) && _initialBalance != 0) {
       _mint(_initialHolder, _initialBalance);
     }