+++ contracts/rUMB.sol
@@ -6,14 +6,13 @@ import "./lib/Strings.sol";
 // Inheritance
 import "./interfaces/SwappableToken.sol";
 import "./interfaces/MintableToken.sol";
-import "./interfaces/BurnableToken.sol";
 
 
 /// @title   Umbrella Rewards contract
 /// @author  umb.network
 /// @notice  This is reward UMB token (rUMB)
 /// @dev     Rewards tokens are used for farming and other rewards distributions.
-contract rUMB is BurnableToken, MintableToken, SwappableToken {
+contract rUMB is MintableToken, SwappableToken {
   using Strings for string;
 
   // ========== STATE VARIABLES ========== //
@@ -21,12 +20,11 @@ contract rUMB is BurnableToken, MintableToken, SwappableToken {
   // ========== CONSTRUCTOR ========== //
 
   constructor (
+    address _multiSig,
     address _initialHolder,
     uint _initialBalance,
-    address _multiSig,
     uint256 _maxAllowedTotalSupply,
     uint _rewardId,
-    uint _totalAmountToBeSwapped,
     uint _swapDuration,
     string memory _name,
     string memory _symbol
@@ -36,8 +34,8 @@ contract rUMB is BurnableToken, MintableToken, SwappableToken {
     _name.appendString(' Reward Token #').appendNumber(_rewardId),
     string('r').appendString(_symbol).appendNumber(_rewardId)
   )
-  MaxAllowedSupply(_maxAllowedTotalSupply)
-  SwappableToken(_totalAmountToBeSwapped, _swapDuration) {
+  MintableToken(_maxAllowedTotalSupply)
+  SwappableToken(_maxAllowedTotalSupply, _swapDuration) {
     if (_initialHolder != address(0) && _initialBalance != 0) {
       _mint(_initialHolder, _initialBalance);
     }